# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления задачами и тегами",
        "title": "Task Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@taskmanager.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/tags": {
            "get": {
                "description": "Получение списка всех тегов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Список тегов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.Tag"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создание нового тега",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Создать тег",
                "parameters": [
                    {
                        "description": "Данные тега",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.CreateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.Tag"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/tags/{id}": {
            "get": {
                "description": "Получение тега по идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Получить тег по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID тега",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.Tag"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление тега по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Удалить тег",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID тега",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/tasks/{taskId}/tags": {
            "get": {
                "description": "Получение списка тегов, привязанных к задаче",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Получить теги задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.Tag"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{taskId}/tags/{tagId}": {
            "post": {
                "description": "Привязка тега к задаче",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Добавить тег к задаче",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID тега",
                        "name": "tagId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные тега (опционально)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.CreateTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Отвязка тега от задачи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Удалить тег из задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID тега",
                        "name": "tagId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID задачи или формат даты",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Переопределение не найдено",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Vladroon22_TaskTracker_internal_models.BadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Not found"
                }
            }
        },
        "github_com_Vladroon22_TaskTracker_internal_models.CreateTagRequest": {
            "type": "object",
            "required": [
                "names"
            ],
            "properties": {
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_Vladroon22_TaskTracker_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "github_com_Vladroon22_TaskTracker_internal_models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
        "github_com_Vladroon22_TaskTracker_internal_models.ServerErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
        "github_com_Vladroon22_TaskTracker_internal_models.Tag": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
...
}